refactor(mbk/pages): replace stacked ternary chains with useXxxMode + flat switch bodies
Implements jkwon-claude-config #92. Every `pages/*.tsx` file that had a ternary
chain of ≥ 3 levels now delegates its render logic to a discriminated union hook
and a dedicated body component, following the DocumentViewer pattern.

Pages refactored (9):
  - Vendors           → useVendorsListMode + VendorsListBody
  - Listings          → useListingsListMode + ListingsListBody
  - Tenants           → useTenantsListMode + TenantsListBody
  - Leases            → useLeasesListMode + LeasesListBody
  - LeaseTemplates    → useLeaseTemplatesListMode + LeaseTemplatesListBody
  - InsurancePolicies → useInsurancePoliciesListMode + InsurancePoliciesListBody
  - InsurancePolicyDetail → useInsurancePolicyDetailMode + InsurancePolicyDetailBody
  - Inquiries         → useInquiriesListMode + InquiriesListBody
  - InquiryDetail     → useInquiryDetailMode + InquiryDetailBody

New discriminated union types (9):
  shared/types/vendor/vendors-list-mode.ts
  shared/types/listing/listings-list-mode.ts
  shared/types/applicant/tenants-list-mode.ts
  shared/types/lease/leases-list-mode.ts
  shared/types/lease/lease-templates-list-mode.ts
  shared/types/insurance/insurance-policies-list-mode.ts
  shared/types/insurance/insurance-policy-detail-mode.ts
  shared/types/inquiry/inquiries-list-mode.ts
  shared/types/inquiry/inquiry-detail-mode.ts

Unit tests: usePageListModes.test.ts covers all 9 hooks (38 test cases).
test-map.json updated for vendors, listings, tenants, leases, insurance, inquiries domains.

No behavioral changes — pure structural refactor.
