+++ apps/mybookkeeper/frontend/src/app/features/leases/LeasesListBody.tsx
@@ -1,0 +1,115 @@
+import { Link } from "react-router-dom";
+import type { LeasesListMode } from "@/shared/types/lease/leases-list-mode";
+import type { SignedLeaseSummary } from "@/shared/types/lease/signed-lease-summary";
+import EmptyState from "@/shared/components/ui/EmptyState";
+import LeasesListSkeleton from "./LeasesListSkeleton";
+import SignedLeaseStatusBadge from "./SignedLeaseStatusBadge";
+
+export interface LeasesListBodyProps {
+  mode: LeasesListMode;
+  leases: SignedLeaseSummary[];
+}
+
+function formatDate(iso: string): string {
+  const d = new Date(iso);
+  return Number.isNaN(d.getTime()) ? "—" : d.toLocaleDateString();
+}
+
+export default function LeasesListBody({ mode, leases }: LeasesListBodyProps) {
+  switch (mode) {
+    case "loading":
+      return <LeasesListSkeleton />;
+    case "empty":
+      return (
+        <EmptyState message="No leases yet — generate one from a template or import an already-signed PDF using the button above." />
+      );
+    case "list":
+      return (
+        <>
+          {/* Desktop table */}
+          <div
+            className="hidden md:block border rounded-lg overflow-hidden"
+            data-testid="leases-table"
+          >
+            <table className="w-full text-sm">
+              <thead className="bg-muted/50 text-left text-xs uppercase text-muted-foreground">
+                <tr>
+                  <th className="px-4 py-2 font-medium">Lease</th>
+                  <th className="px-4 py-2 font-medium">Tenant</th>
+                  <th className="px-4 py-2 font-medium">Term</th>
+                  <th className="px-4 py-2 font-medium">Generated</th>
+                  <th className="px-4 py-2 font-medium">Signed</th>
+                  <th className="px-4 py-2 font-medium">Status</th>
+                </tr>
+              </thead>
+              <tbody>
+                {leases.map((lease) => (
+                  <tr
+                    key={lease.id}
+                    className="border-t hover:bg-muted/40 cursor-pointer"
+                    data-testid={`lease-row-${lease.id}`}
+                  >
+                    <td className="px-4 py-2">
+                      <Link
+                        to={`/leases/${lease.id}`}
+                        className="text-primary hover:underline"
+                      >
+                        Lease {lease.id.slice(0, 8)}
+                      </Link>
+                    </td>
+                    <td className="px-4 py-2">
+                      {lease.applicant_legal_name ? (
+                        <Link
+                          to={`/applicants/${lease.applicant_id}`}
+                          className="text-primary hover:underline"
+                          data-testid={`lease-tenant-link-${lease.id}`}
+                        >
+                          {lease.applicant_legal_name}
+                        </Link>
+                      ) : (
+                        <span className="text-muted-foreground">—</span>
+                      )}
+                    </td>
+                    <td className="px-4 py-2 text-muted-foreground">
+                      {lease.starts_on ?? "—"} → {lease.ends_on ?? "—"}
+                    </td>
+                    <td className="px-4 py-2 text-muted-foreground">
+                      {lease.generated_at ? formatDate(lease.generated_at) : "—"}
+                    </td>
+                    <td className="px-4 py-2 text-muted-foreground">
+                      {lease.signed_at ? formatDate(lease.signed_at) : "—"}
+                    </td>
+                    <td className="px-4 py-2">
+                      <SignedLeaseStatusBadge status={lease.status} />
+                    </td>
+                  </tr>
+                ))}
+              </tbody>
+            </table>
+          </div>
+
+          {/* Mobile cards */}
+          <ul className="md:hidden space-y-3" data-testid="leases-mobile">
+            {leases.map((lease) => (
+              <li key={lease.id}>
+                <Link
+                  to={`/leases/${lease.id}`}
+                  className="block border rounded-lg p-3 hover:bg-muted/40"
+                >
+                  <div className="flex items-center justify-between">
+                    <span className="font-medium text-sm">
+                      Lease {lease.id.slice(0, 8)}
+                    </span>
+                    <SignedLeaseStatusBadge status={lease.status} />
+                  </div>
+                  <p className="text-xs text-muted-foreground mt-1">
+                    {lease.starts_on ?? "—"} → {lease.ends_on ?? "—"}
+                  </p>
+                </Link>
+              </li>
+            ))}
+          </ul>
+        </>
+      );
+  }
+}
