+++ apps/mybookkeeper/frontend/src/app/pages/Inquiries.tsx
@@ -2,7 +2,6 @@ import { useMemo, useState } from "react";
 import { FileText, Plus } from "lucide-react";
 import { Link, useSearchParams } from "react-router-dom";
 import SectionHeader from "@/shared/components/ui/SectionHeader";
-import EmptyState from "@/shared/components/ui/EmptyState";
 import AlertBox from "@/shared/components/ui/AlertBox";
 import LoadingButton from "@/shared/components/ui/LoadingButton";
 import { useGetInquiriesQuery } from "@/shared/store/inquiriesApi";
@@ -10,12 +9,11 @@ import { useGetListingsQuery } from "@/shared/store/listingsApi";
 import { INQUIRY_PAGE_SIZE, INQUIRY_STAGES } from "@/shared/lib/inquiry-labels";
 import type { InquirySpamStatus } from "@/shared/types/inquiry/inquiry-spam-status";
 import type { InquiryStage } from "@/shared/types/inquiry/inquiry-stage";
-import InquiriesSkeleton from "@/app/features/inquiries/InquiriesSkeleton";
 import InquiryStageFilter from "@/app/features/inquiries/InquiryStageFilter";
 import InquirySpamTabFilter from "@/app/features/inquiries/InquirySpamTabFilter";
-import InquiryCard from "@/app/features/inquiries/InquiryCard";
-import InquiryRow from "@/app/features/inquiries/InquiryRow";
 import InquiryForm from "@/app/features/inquiries/InquiryForm";
+import { useInquiriesListMode } from "@/app/features/inquiries/useInquiriesListMode";
+import InquiriesListBody from "@/app/features/inquiries/InquiriesListBody";
 
 const STAGE_PARAM = "stage";
 const SPAM_PARAM = "spam";
@@ -116,8 +114,9 @@ export default function Inquiries() {
   }
 
   const showStageBadge = stage === null;
-
   const isFiltered = stage !== null;
+
+  const mode = useInquiriesListMode({ isLoading, isError, inquiryCount: inquiries.length });
 
   return (
     <main className="p-4 sm:p-8 space-y-6">
@@ -165,75 +164,16 @@ export default function Inquiries() {
         </AlertBox>
       ) : null}
 
-      {isLoading ? (
-        <InquiriesSkeleton />
-      ) : inquiries.length === 0 && !isError ? (
-        <EmptyState
-          message={
-            isFiltered
-              ? "No inquiries in this stage. Try a different filter."
-              : "No inquiries yet. They'll land here when guests reach out via Furnished Finder, TNH, or directly. Use \"New inquiry\" to log one manually."
-          }
-        />
-      ) : (
-        <>
-          {/* Mobile: cards */}
-          <ul className="md:hidden space-y-3" data-testid="inquiries-mobile">
-            {inquiries.map((inquiry) => (
-              <li key={inquiry.id}>
-                <InquiryCard
-                  inquiry={inquiry}
-                  listingTitle={inquiry.listing_id ? listingTitleById.get(inquiry.listing_id) ?? null : null}
-                  showStageBadge={showStageBadge}
-                />
-              </li>
-            ))}
-          </ul>
-
-          {/* Desktop: table */}
-          <div
-            className="hidden md:block border rounded-lg overflow-hidden"
-            data-testid="inquiries-desktop"
-          >
-            <table className="w-full text-sm">
-              <thead className="bg-muted/50 text-left text-xs uppercase text-muted-foreground">
-                <tr>
-                  <th className="px-4 py-2 font-medium">Inquirer</th>
-                  <th className="px-4 py-2 font-medium">Source</th>
-                  <th className="px-4 py-2 font-medium">Desired Dates</th>
-                  <th className="px-4 py-2 font-medium">Employer</th>
-                  <th className="px-4 py-2 font-medium">Listing</th>
-                  <th className="px-4 py-2 font-medium">Received</th>
-                  <th className="px-4 py-2 font-medium">Stage</th>
-                  <th className="px-4 py-2 font-medium">Quality</th>
-                </tr>
-              </thead>
-              <tbody>
-                {inquiries.map((inquiry) => (
-                  <InquiryRow
-                    key={inquiry.id}
-                    inquiry={inquiry}
-                    listingTitle={inquiry.listing_id ? listingTitleById.get(inquiry.listing_id) ?? null : null}
-                  />
-                ))}
-              </tbody>
-            </table>
-          </div>
-
-          {hasMore ? (
-            <div className="flex justify-center">
-              <LoadingButton
-                variant="secondary"
-                onClick={handleLoadMore}
-                isLoading={isFetching}
-                loadingText="Loading..."
-              >
-                Load more
-              </LoadingButton>
-            </div>
-          ) : null}
-        </>
-      )}
+      <InquiriesListBody
+        mode={mode}
+        inquiries={inquiries}
+        isFiltered={isFiltered}
+        showStageBadge={showStageBadge}
+        listingTitleById={listingTitleById}
+        hasMore={hasMore}
+        isFetching={isFetching}
+        onLoadMore={handleLoadMore}
+      />
 
       {showForm ? (
         <InquiryForm
