+++ apps/mybookkeeper/frontend/src/app/pages/InsurancePolicies.tsx
@@ -1,10 +1,10 @@
 import { useState } from "react";
-import { Link } from "react-router-dom";
 import { addDays, format } from "date-fns";
 import { useGetInsurancePoliciesQuery } from "@/shared/store/insurancePoliciesApi";
 import SectionHeader from "@/shared/components/ui/SectionHeader";
 import AlertBox from "@/shared/components/ui/AlertBox";
-import InsuranceExpirationBadge from "@/app/features/insurance/InsuranceExpirationBadge";
+import { useInsurancePoliciesListMode } from "@/app/features/insurance/useInsurancePoliciesListMode";
+import InsurancePoliciesListBody from "@/app/features/insurance/InsurancePoliciesListBody";
 
 /**
  * All-policies view: lists insurance policies across all listings.
@@ -22,6 +22,7 @@ export default function InsurancePolicies() {
   );
 
   const policies = data?.items ?? [];
+  const mode = useInsurancePoliciesListMode({ isLoading, policyCount: policies.length });
 
   return (
     <main className="p-4 sm:p-8 space-y-6 max-w-3xl">
@@ -57,56 +58,11 @@ export default function InsurancePolicies() {
         </label>
       </div>
 
-      {isLoading ? (
-        <div
-          className="space-y-2 animate-pulse"
-          aria-busy="true"
-          data-testid="insurance-policies-loading"
-        >
-          {[1, 2, 3].map((i) => (
-            <div key={i} className="border rounded-lg p-4 space-y-2">
-              <div className="h-4 bg-muted rounded w-1/2" />
-              <div className="h-4 bg-muted rounded w-1/3" />
-            </div>
-          ))}
-        </div>
-      ) : policies.length === 0 ? (
-        <p
-          className="text-sm text-muted-foreground"
-          data-testid="insurance-policies-empty"
-        >
-          {showExpiringSoon
-            ? "No policies expiring within 30 days."
-            : "No policies on this listing yet — add one to track coverage and expiration."}
-        </p>
-      ) : (
-        <ul className="space-y-2" data-testid="insurance-policies-list">
-          {policies.map((policy) => (
-            <li key={policy.id} className="border rounded-lg px-4 py-3 text-sm">
-              <div className="flex items-start justify-between gap-3">
-                <div className="min-w-0">
-                  <Link
-                    to={`/insurance-policies/${policy.id}`}
-                    className="font-medium text-primary hover:underline block truncate"
-                    data-testid={`insurance-policy-item-${policy.id}`}
-                  >
-                    {policy.policy_name}
-                  </Link>
-                  {policy.carrier ? (
-                    <p className="text-xs text-muted-foreground mt-0.5">{policy.carrier}</p>
-                  ) : null}
-                  {policy.expiration_date ? (
-                    <p className="text-xs text-muted-foreground mt-0.5">
-                      Expires {format(new Date(policy.expiration_date + "T00:00:00"), "MMM d, yyyy")}
-                    </p>
-                  ) : null}
-                </div>
-                <InsuranceExpirationBadge expirationDate={policy.expiration_date} />
-              </div>
-            </li>
-          ))}
-        </ul>
-      )}
+      <InsurancePoliciesListBody
+        mode={mode}
+        policies={policies}
+        showExpiringSoon={showExpiringSoon}
+      />
     </main>
   );
 }
