+++ apps/mybookkeeper/frontend/src/app/pages/Tenants.tsx
@@ -1,13 +1,11 @@
 import { useMemo, useState } from "react";
 import SectionHeader from "@/shared/components/ui/SectionHeader";
-import EmptyState from "@/shared/components/ui/EmptyState";
 import AlertBox from "@/shared/components/ui/AlertBox";
 import LoadingButton from "@/shared/components/ui/LoadingButton";
 import { useGetTenantsQuery } from "@/shared/store/applicantsApi";
 import { APPLICANT_PAGE_SIZE } from "@/shared/lib/applicant-labels";
-import TenantsListSkeleton from "@/app/features/tenants/TenantsListSkeleton";
-import TenantCard from "@/app/features/tenants/TenantCard";
-import TenantRow from "@/app/features/tenants/TenantRow";
+import { useTenantsListMode } from "@/app/features/tenants/useTenantsListMode";
+import TenantsListBody from "@/app/features/tenants/TenantsListBody";
 
 export default function Tenants() {
   const [includeEnded, setIncludeEnded] = useState(false);
@@ -27,6 +25,8 @@ export default function Tenants() {
 
   const tenants = data?.items ?? [];
   const hasMore = data?.has_more ?? false;
+
+  const mode = useTenantsListMode({ isLoading, isError, tenantCount: tenants.length });
 
   function handleToggleEnded() {
     setIncludeEnded((prev) => !prev);
@@ -72,63 +72,14 @@ export default function Tenants() {
         </AlertBox>
       ) : null}
 
-      {isLoading ? (
-        <TenantsListSkeleton />
-      ) : tenants.length === 0 && !isError ? (
-        <EmptyState
-          message={
-            includeEnded
-              ? "No tenants found — active or ended."
-              : "No active tenants. When you mark an applicant as 'Lease signed', they'll appear here."
-          }
-        />
-      ) : (
-        <>
-          {/* Mobile: cards */}
-          <ul className="md:hidden space-y-3" data-testid="tenants-mobile">
-            {tenants.map((tenant) => (
-              <li key={tenant.id}>
-                <TenantCard tenant={tenant} />
-              </li>
-            ))}
-          </ul>
-
-          {/* Desktop: table */}
-          <div
-            className="hidden md:block border rounded-lg overflow-hidden"
-            data-testid="tenants-desktop"
-          >
-            <table className="w-full text-sm">
-              <thead className="bg-muted/50 text-left text-xs uppercase text-muted-foreground">
-                <tr>
-                  <th className="px-4 py-2 font-medium">Name</th>
-                  <th className="px-4 py-2 font-medium">Contract Dates</th>
-                  <th className="px-4 py-2 font-medium">Since</th>
-                  <th className="px-4 py-2 font-medium">Status</th>
-                </tr>
-              </thead>
-              <tbody>
-                {tenants.map((tenant) => (
-                  <TenantRow key={tenant.id} tenant={tenant} />
-                ))}
-              </tbody>
-            </table>
-          </div>
-
-          {hasMore ? (
-            <div className="flex justify-center">
-              <LoadingButton
-                variant="secondary"
-                onClick={handleLoadMore}
-                isLoading={isFetching}
-                loadingText="Loading..."
-              >
-                Load more
-              </LoadingButton>
-            </div>
-          ) : null}
-        </>
-      )}
+      <TenantsListBody
+        mode={mode}
+        tenants={tenants}
+        includeEnded={includeEnded}
+        hasMore={hasMore}
+        isFetching={isFetching}
+        onLoadMore={handleLoadMore}
+      />
     </main>
   );
 }
