+++ apps/mybookkeeper/frontend/src/shared/types/applicant/tenants-list-mode.ts
@@ -1,0 +1,5 @@
+/**
+ * Discriminated union for what the Tenants list body should render.
+ * Replaces the isLoading / empty / list ternary chain with a flat switch.
+ */
+export type TenantsListMode = "loading" | "empty" | "list";
