+++ apps/mybookkeeper/frontend/src/shared/types/insurance/insurance-policy-detail-mode.ts
@@ -1,0 +1,5 @@
+/**
+ * Discriminated union for what the InsurancePolicyDetail body should render.
+ * Replaces the nested isLoading / isError / content ternary chain with a flat switch.
+ */
+export type InsurancePolicyDetailMode = "loading" | "content";
