+++ apps/mybookkeeper/frontend/src/shared/types/listing/listings-list-mode.ts
@@ -1,0 +1,5 @@
+/**
+ * Discriminated union for what the Listings list body should render.
+ * Replaces the isLoading / empty / list ternary chain with a flat switch.
+ */
+export type ListingsListMode = "loading" | "empty" | "list";
